+++ CHANGELOG.md
@@ -7,6 +7,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/), and this
 
 ## Unreleased
 
+### Fixed
+
+* Fix Fix missing http_proxy in managment handler
+
 ### Added
 
 * Add informative warning on dangling plugin symlink